pyproject.toml
[tool.poetry]
name = "personik"
version = "0.1.0"
description = ""
authors = ["Maksim <temkinmaksim17@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
bs4 = "^0.0.1"
requests = "^2.28.2"
python-time = "^0.3.0"
httpx = "^0.23.3"
pytest = "^7.3.0"
google-serp-api = "^1.0.3"
google-search-results = "^2.4.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
